refactor(mainnet): add pallet_multisig to utility
