+++ runtime/mainnet/src/config/mod.rs
@@ -1,2 +1,3 @@
 mod proxy;
+mod utility;
 pub mod xcm;
